Port recent libgit2 work to windows

+++ foreign/windows/libgit2/include/git2/sys/index.h
@@ -42,7 +42,7 @@ typedef struct git_index_reuc_entry {
  * @param index an existing index object
  * @return integer of count of current filename conflict entries
  */
-GIT_EXTERN(unsigned int) git_index_name_entrycount(git_index *index);
+GIT_EXTERN(size_t) git_index_name_entrycount(git_index *index);
 
 /**
  * Get a filename conflict entry from the index.
@@ -72,7 +72,6 @@ GIT_EXTERN(int) git_index_name_add(git_index *index,
  * Remove all filename conflict entries.
  *
  * @param index an existing index object
- * @return 0 or an error code
  */
 GIT_EXTERN(void) git_index_name_clear(git_index *index);
 
@@ -91,7 +90,7 @@ GIT_EXTERN(void) git_index_name_clear(git_index *index);
  * @param index an existing index object
  * @return integer of count of current resolve undo entries
  */
-GIT_EXTERN(unsigned int) git_index_reuc_entrycount(git_index *index);
+GIT_EXTERN(size_t) git_index_reuc_entrycount(git_index *index);
 
 /**
  * Finds the resolve undo entry that points to the given path in the Git
@@ -168,7 +167,6 @@ GIT_EXTERN(int) git_index_reuc_remove(git_index *index, size_t n);
  * Remove all resolve undo entries from the index
  *
  * @param index an existing index object
- * @return 0 or an error code
  */
 GIT_EXTERN(void) git_index_reuc_clear(git_index *index);
